scripts/deploy-cip-core.sh: Check for the existence of the broken swu file before uploading
Commit e1c4b6fa (ci: Build and deploy kernel-panic update artifact) introduced
deploying a broken swu file for kernel panic test in the CI. But not all CI jobs
builds a broken swu file (ex: Reproducible build tests)

Add a check for the existence of the broken swu file before attempting an upload

+++ scripts/deploy-cip-core.sh
@@ -49,7 +49,9 @@ if [ -f "${BASE_PATH}.wic" ]; then
 	echo "Uploading artifacts..."
 	if [ "$DEPLOY" = "swus" ]; then
 		aws s3 cp --no-progress --acl public-read "${BASE_PATH}.swu" "${S3_TARGET}"
-		aws s3 cp --no-progress --acl public-read "${BASE_PATH}-broken.swu" "${S3_TARGET}"
+		if [ -f "${BASE_PATH}-broken.swu" ]; then
+			aws s3 cp --no-progress --acl public-read "${BASE_PATH}-broken.swu" "${S3_TARGET}"
+		fi
 	elif [ "$DEPLOY" = "wic-partitions" ]; then
 		# deploy individual wic partitions, helpful for RB tests
 		cd build/tmp/deploy/images/"$TARGET"
